the camera goes to f64, and the wall moves nine orders of magnitude
`pan_x`/`pan_y` are f64. They are the only ABSOLUTE positions among
the view fields, and an absolute position is the one thing a deep
zoom destroys: one f32 ulp near 0.27 is 3e-8, wider than the whole
viewport past a zoom of about 1e8, so the centre could not be
ADDRESSED however well the render was sampled.

Everything downstream already had the precision to use it. The
forced-prefix packing has subtracted the view centre in f64 since the
relative-coordinates commit; the input was simply quantised before it
got there. Measured on the reported fern, structure now resolves to
**1e17** and the frame is empty by 1e18 -- which is f64's own
resolution at that centre meeting the view half-width, so the new
wall is the representation rather than an accident.

`zoom` stays f32 deliberately. It is a SCALE, so only relative
precision matters and f32 has seven digits of that at any magnitude;
widening it would be cost without effect.

No config migration and no version bump: a JSON number parses into
f64 whatever it was written as, and an older build reading a new file
degrades to f32 gracefully.

Shape of the change: `ConfigValue::Vec2` becomes f64 (the pan is its
only user -- the enum's own comment said so), plus a new `Double`
variant for `PanX`/`PanY` so the several hundred genuinely-f32
parameters keep `Float`. Animation reads `Double` first and widens a
`Float`, so a track written before this still drives the pan. About a
hundred boundary casts: `as f32` where the pan meets the GPU uniform
or the 3D shadow placement, `as f64` where it meets a screen delta.
Nothing in the shaders changes, which is right -- under relative
packing the kernel never reads the pan at all.

**There is deliberately no `From<f64>` for ConfigValue.** Adding one
silently retypes every `2.0.into()` in the codebase from `Float` to
`Double`, because an untyped float literal infers f64 -- and the
failure lands at RUNTIME as a type mismatch, not at compile time. Four
config tests caught it; the two paths that want a `Double` name it.

The gate's fixture was worth as much as the change. Its centre is now
SOLVED as the fixed point of the map as stored -- in f32, widened --
because a centre computed from the decimal coefficients sits 6.4e-9
away, which is inside the viewport at 1e8 and outside at 1e9. That
made a correct implementation read as a precision failure, and cost
an hour before I checked the fixture instead of the code.

330/330 visual, all release gates, 1274 unit tests, 7 cylinder GPU
gates.

+++ src/app/fly_camera.rs
@@ -478,8 +478,8 @@ impl App {
             // factor.
             let m_old_nr = CameraMatrix::build(pitch_old, yaw_old, 0.0);
             let m_new_nr = CameraMatrix::build(pitch_new, yaw_new, 0.0);
-            let off_old = m_old_nr.world_offset_for_camera_xy(pan_x, pan_y);
-            let off_new = m_new_nr.world_offset_for_camera_xy(pan_x, pan_y);
+            let off_old = m_old_nr.world_offset_for_camera_xy(pan_x as f32, pan_y as f32);
+            let off_new = m_new_nr.world_offset_for_camera_xy(pan_x as f32, pan_y as f32);
             changes.push((ConfigPath::CameraX, (cam_x + off_old[0] - off_new[0]).into()));
             changes.push((ConfigPath::CameraY, (cam_y + off_old[1] - off_new[1]).into()));
             changes.push((ConfigPath::CameraZ, (cam_z + off_old[2] - off_new[2]).into()));
